Minor refactor Cleanup after merge
