+++ lib/src/bloc/auth/auth.dart
