+++ lib/src/bloc/auth/auth_event.dart
