+++ lib/src/bloc/auth/auth_state.dart
