+++ lib/src/bloc/login/login_event.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
-
 import 'package:meta/meta.dart';
 import 'package:equatable/equatable.dart';
+
 
 abstract class LoginEvent extends Equatable {
   LoginEvent([List props = const []]) : super(props);
