+++ lib/src/bloc/nav/navbar_bloc.dart
