+++ lib/src/bloc/nav/navbar_event.dart
