+++ lib/src/bloc/nav/navbar_state.dart
